requirements.txt
langchain>=0.1.4
langchain-openai>=0.0.5
langchain-community>=0.0.16
langchain-core>=0.1.17
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.32.0
biopython==1.83
tiktoken>=0.7.0
